feat: recognize and use over sized allocations
Allocators are allowed to return a larger memory chunk than was asked
for. If the amount extra is large enough, then the hash table can use
the extra space. The Global allocator will not hit this path, because
it won't over-size enough to matter, but custom allocators may. An
example of an allocator which allocates full system pages is included
in the test suite (UNIX only because it uses `mmap`).